Multiple minor changes
Added Average Distance Divergence, and attempted to fixthe  ID problems

+++ include/distributions.h
@@ -4,8 +4,8 @@
 #include <vector>
 
 	unsigned uniform_distribution(unsigned lo, unsigned hi);
-	int normal_distribution(int mean, int variance);
+	double normal_distribution(int mean, int variance);
 	
-	std::vector<int> normal_vector(int mean, int variance, unsigned v_size);
+	std::vector<double> normal_vector(int mean, int variance, unsigned v_size);
 
 #endif
